+++ src/main/java/DAO/BookDao.java
@@ -1,5 +1,7 @@
 package DAO;
 
+import com.opencsv.CSVReader;
+import com.opencsv.exceptions.CsvException;
 import connection.ConnectionManager;
 import exceptions.DuplicateObjectException;
 import objects.Author;
@@ -9,9 +11,11 @@ import validating.BookValidator;
 import validating.Validator;
 
 import java.io.FileInputStream;
+import java.io.FileReader;
 import java.io.IOException;
 import java.sql.Date;
 import java.sql.*;
+import java.text.SimpleDateFormat;
 import java.util.*;
 
 import static DAO.Constants.SQL_PATH;
@@ -22,12 +26,16 @@ public class BookDao implements DefaultDao<Book, Integer> {
     private final AuthorDao authorDAO;
     private final BookCategoryDao categoryDAO;
 
+    private Validator<Book> validator;
+
     public BookDao() throws IOException {
         QUERIES = new Properties();
         QUERIES.load(new FileInputStream(SQL_PATH));
 
         authorDAO = new AuthorDao();
         categoryDAO = new BookCategoryDao();
+
+        validator = new BookValidator();
     }
 
     @Override
@@ -69,7 +77,6 @@ public class BookDao implements DefaultDao<Book, Integer> {
                                                                                                                             obj.getTitle(),
                                                                                                                             obj.getRecommendedPrice(),
                                                                                                                             obj.getComments());
-                Validator<Book> validator = new BookValidator();
                 if (validator.isValid(obj)) {
                     return statement.executeUpdate() > 0;
                 }
@@ -93,7 +100,6 @@ public class BookDao implements DefaultDao<Book, Integer> {
                                                                                                                                  obj.getRecommendedPrice(),
                                                                                                                                  obj.getComments(),
                                                                                                                                  obj.getId())) {
-            Validator<Book> validator = new BookValidator();
             if (validator.isValid(obj)) {
                 return preparedStatement.executeUpdate() > 0;
             }
@@ -105,11 +111,14 @@ public class BookDao implements DefaultDao<Book, Integer> {
 
     @Override
     public Boolean delete(Book obj) {
-        try (PreparedStatement statement = ConnectionManager.prepareStatement(QUERIES.getProperty("deleteByID.query").replaceAll("TABLE_NAME", TABLE), obj.getId())) {
-            return statement.executeUpdate() > 0;
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
+        if (obj != null) {
+            try (PreparedStatement statement = ConnectionManager.prepareStatement(QUERIES.getProperty("deleteByID.query").replaceAll("TABLE_NAME", TABLE), obj.getId())) {
+                return statement.executeUpdate() > 0;
+            } catch (SQLException e) {
+                throw new RuntimeException(e);
+            }
         }
+        return false;
     }
 
     @Override
@@ -125,5 +134,44 @@ public class BookDao implements DefaultDao<Book, Integer> {
         String title = resultSet.getString("title");
 
         return new Book(id, author, category, comments, acquiringDate, isbn, publicationDate, recommendedPrice, title);
+    }
+
+    @Override
+    public Book convertObject(String[] objectLine) {
+        int id = Integer.parseInt(objectLine[0]);
+        Author author = authorDAO.findByKey(Integer.parseInt(objectLine[1])).orElse(null);
+        BookCategory category = categoryDAO.findByKey(Integer.parseInt(objectLine[2])).orElse(null);
+        String isbn = objectLine[3];
+        Date publicationDate = Date.valueOf(objectLine[4]);
+        Date acquiringDate = Date.valueOf(objectLine[5]);
+        String title = objectLine[6];
+        Double recommendedPrice = Double.parseDouble(objectLine[7]);
+        String comments = objectLine[8];
+
+        return new Book(id, author, category, comments, acquiringDate, isbn, publicationDate, recommendedPrice, title);
+    }
+
+    @Override
+    public String[] convertObjectForCsvExport(Book object) {
+        String pattern = "yyyy-M-d";
+        SimpleDateFormat dateFormat = new SimpleDateFormat(pattern);
+
+        String[] line = {
+                object.getId().toString(),
+                object.getAuthor().getId().toString(),
+                object.getCategory().getCode().toString(),
+                object.getIsbn(),
+                dateFormat.format(object.getPublicationDate()),
+                dateFormat.format(object.getDateAcquired()),
+                object.getTitle(),
+                object.getRecommendedPrice().toString(),
+                object.getComments()
+        };
+
+        return line;
+    }
+
+    public String[] getHeaderLine(){
+        return new String[]{"id", "author_id", "book_category_code", "isbn", "publication_date", "date_aquired", "title", "recommended_price", "comments"};
     }
 }
